Sang computer conflict fix

+++ fashionShop-BE/src/main/java/com/fashionshopbe/repository/ProductImageRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
-    List<ProductImage> findByProductId(int productId);
+    List<ProductImage> findByProductId(long productId);
 }
